python/requirements.txt
pillow~=10.3.0
ultralytics==8.2.28
